+++ packages/product-catalogue/src/ducks/productCatalogue/index.ts
