+++ packages/product-catalogue/src/ducks/productCatalogue/tests/fixtures.ts
